Code refactoring

+++ 42cursus/Libft/ft_strlen.c
@@ -3,17 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: alperrot <alperrot@student.42angouleme.fr  +#+  +:+       +#+        */
+/*   By: alperrot <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 16:14:06 by alperrot          #+#    #+#             */
-/*   Updated: 2023/11/07 11:45:49 by alperrot         ###   ########.fr       */
+/*   Updated: 2023/11/11 15:01:55 by alperrot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
 #include "libft.h"
 
 size_t	ft_strlen(const char *str)
 {
-	long unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
